Fix invalid output fault on IO extender bistable relays
Relays 1 & 2 on the IO extender would give an output fault when
the state changes rapidly (only relevent with type momentary)

This was caused by the thread that is watching the feedback pin and the desired state.
The desired state changed during waiting and differed from what the feedback pin read.

The fix is to properly cancel the waiting thread and
synchronize access to the shared resources.

https://github.com/victronenergy/venus/issues/1529